# dashboard/tmp/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: Scheduler Public API (v1)
  version: 1.3.0
servers:
  - url: 'https://scheduler.example.com'
security:
  - bearerAuth: []
...
paths:
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Exchange username/password for a JWT
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/runs:
    get:
      tags:
        - Runs
      summary: List runs (paginated)
      operationId: listRuns
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RunStatus'
        - name: clientId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
    post:
      tags:
        - Runs
      summary: Start a run using the in-memory workflow snapshot
      operationId: startRun
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunStartRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunRef'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  '/api/v1/runs/{runId}':
    get:
      tags:
        - Runs
      summary: Get run summary
      operationId: getRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          $ref: '#/components/responses/NotFound'
  '/api/v1/runs/{runId}/cancel':
    post:
      tags:
        - Runs
      summary: Cancel a run
      operationId: cancelRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunRef'
        '404':
          $ref: '#/components/responses/NotFound'
  '/api/v1/runs/{runId}/definition':
    get:
      tags:
        - Runs
      summary: Get the immutable workflow snapshot used by this run
      operationId: getRunDefinition
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: OK
          headers:
            ETag:
              description: Weak ETag based on definitionHash
              schema:
                type: string
                example: 'W/"sha256:9e22a4..."'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '404':
          $ref: '#/components/responses/NotFound'
...
  '/api/v1/packages/{packageName}':
    get:
      tags:
        - Packages
      summary: Get package detail
      operationId: getPackage
      parameters:
        - name: packageName
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: query
          required: false
          description: Specific package version to retrieve. Defaults to the latest available version.
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/users:
    get:
      tags:
        - Users
      summary: List users and their roles
      operationId: listUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
    post:
      tags:
        - Users
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
  '/api/v1/users/{userId}/password':
    post:
      tags:
        - Users
      summary: Reset user password
      operationId: resetUserPassword
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetUserPasswordRequest'
      responses:
        '204':
          description: Password updated
  '/api/v1/users/{userId}/roles':
    post:
      tags:
        - Users
      summary: Assign role to user
      operationId: addUserRole
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleRequest'
      responses:
        '204':
          description: Role assigned
  '/api/v1/users/{userId}/roles/{role}':
    delete:
      tags:
        - Users
      summary: Remove role from user
      operationId: removeUserRole
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: role
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Role removed
  '/api/v1/users/{userId}/status':
    patch:
      tags:
        - Users
      summary: Toggle user active state
      operationId: updateUserStatus
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
      responses:
        '204':
          description: Status updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 64
      description: 'Optional idempotency key for safe retries; if reused with a different body, return 409'
    PageLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    PageCursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 'Conflict (e.g., idempotency-key reuse with different body)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or credentials invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    RunStartRequest:
      type: object
      required:
        - workflow
        - clientId
      properties:
        workflow:
          $ref: '#/components/schemas/Workflow'
        clientId:
          type: string
          description: Caller-provided client instance id
    RunRef:
      type: object
      required:
        - runId
        - status
        - definitionHash
        - clientId
        - createdAt
      properties:
        runId:
          type: string
        status:
          $ref: '#/components/schemas/RunStatus'
        definitionHash:
          type: string
        clientId:
          type: string
        createdAt:
          type: string
          format: date-time
    Run:
      type: object
      required:
        - runId
        - status
        - definitionHash
        - clientId
      properties:
        runId:
          type: string
        status:
          $ref: '#/components/schemas/RunStatus'
        definitionHash:
          type: string
        clientId:
          type: string
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        error:
          $ref: '#/components/schemas/ResultError'
          nullable: true
        artifacts:
          $ref: '#/components/schemas/RunArtifactList'
          nullable: true
        nodes:
          $ref: '#/components/schemas/RunNodesList'
          nullable: true
    RunStatus:
      type: string
      enum:
        - queued
        - running
        - succeeded
        - failed
        - cancelled
    RunList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Run'
        nextCursor:
          type: string
          nullable: true
    RunArtifact:
      type: object
      required:
        - resourceId
        - workerId
        - type
      properties:
        resourceId:
          type: string
        workerId:
          type: string
        type:
          type: string
        sizeBytes:
          type: integer
          minimum: 0
          nullable: true
        inline:
          type: boolean
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
    RunArtifactList:
      type: array
      items:
        $ref: '#/components/schemas/RunArtifact'
...
    RunNodesList:
      type: array
      items:
        $ref: '#/components/schemas/RunNodeStatus'
    Workflow:
      type: object
      additionalProperties: false
      required:
        - id
        - schemaVersion
        - metadata
        - nodes
        - edges
      properties:
        id:
          type: string
          format: uuid
          description: Workflow UUID
        schemaVersion:
          type: string
          description: e.g. "2025-10"
          pattern: '^[0-9]{4}-[0-9]{2}$'
        metadata:
          $ref: '#/components/schemas/WorkflowMetadata'
        nodes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/WorkflowNode'
        edges:
          type: array
          minItems: 0
          items:
            $ref: '#/components/schemas/WorkflowEdge'
        tags:
          type: array
          description: Workflow-level tags.
          items:
            type: string
        previewImage:
          type: string
          description: Base64-encoded preview of the workflow canvas.
          nullable: true
        subgraphs:
          type: array
          description: Reusable workflow fragments (localized snapshots) that container nodes can reference.
          items:
            $ref: '#/components/schemas/WorkflowSubgraph'
...
    WorkflowSubgraph:
      type: object
      additionalProperties: false
      required:
        - id
        - definition
      properties:
        id:
          type: string
          description: Stable identifier referenced by container nodes.
        definition:
          $ref: '#/components/schemas/Workflow'
          description: Localized workflow snapshot backing this subgraph.
        metadata:
          $ref: '#/components/schemas/WorkflowSubgraphMetadata'
...
    AuthLoginRequest:
      type: object
      additionalProperties: false
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    AuthLoginResponse:
      type: object
      additionalProperties: false
      required:
        - accessToken
        - tokenType
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum:
            - Bearer
        expiresIn:
          type: integer
          description: Seconds until the access token expires.
        user:
          $ref: '#/components/schemas/UserSummary'
    UserSummary:
      type: object
      additionalProperties: false
      required:
        - userId
        - username
        - displayName
        - roles
        - isActive
      properties:
        userId:
          type: string
        username:
          type: string
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string
        isActive:
          type: boolean
          description: Whether the account is enabled.
    UserList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
    CreateUserRequest:
      type: object
      additionalProperties: false
      required:
        - username
        - displayName
        - password
      properties:
        username:
          type: string
        displayName:
          type: string
        password:
          type: string
          format: password
        roles:
          type: array
          items:
            type: string
          description: Optional initial roles to assign to the user.
    ResetUserPasswordRequest:
      type: object
      additionalProperties: false
      required:
        - password
      properties:
        password:
          type: string
          format: password
    UserRoleRequest:
      type: object
      additionalProperties: false
      required:
        - role
      properties:
        role:
          type: string
    UpdateUserStatusRequest:
      type: object
      additionalProperties: false
      required:
        - isActive
      properties:
        isActive:
          type: boolean
...
    PackageDetail:
      type: object
      additionalProperties: false
      required:
        - name
        - version
        - manifest
      properties:
        name:
          type: string
        version:
          type: string
        availableVersions:
          type: array
          items:
            type: string
        manifest:
          $ref: '#/components/schemas/PackageManifest'
    PackageManifest:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - name
        - version
        - description
        - adapters
        - python
        - nodes
      properties:
        schemaVersion:
          type: string
          description: Manifest schema version (semver).
          pattern: '^[0-9]+\.[0-9]+\.[0-9]+$'
        name:
          type: string
          description: Package identifier (lowercase with dots/underscores).
          pattern: '^[a-z0-9]+([._-][a-z0-9]+)*$'
        version:
          type: string
          description: Package version in semver format.
          pattern: '^[0-9]+\.[0-9]+\.[0-9]+(?:[-+][0-9A-Za-z\.-]+)?$'
        description:
          type: string
          description: Short summary of the package.
        adapters:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ManifestAdapter'
        python:
          $ref: '#/components/schemas/ManifestPythonConfig'
        nodes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ManifestNode'
        resources:
          type: array
          items:
            $ref: '#/components/schemas/ManifestResource'
        hooks:
          $ref: '#/components/schemas/ManifestHooks'
        signature:
          $ref: '#/components/schemas/ManifestSignature'
...
    WorkflowMetadata:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        environment:
          type: string
        namespace:
          type: string
          description: Logical workflow namespace used for indexing/cross-workflow bindings. Defaults to "default".
        originId:
          type: string
          format: uuid
          description: Identifier linking versions of the same workflow. By default equals the workflow id.
        ownerId:
          type: string
          description: User id that owns the workflow definition.
        ownerName:
          type: string
          description: Display name of the owner if available.
          nullable: true
        createdBy:
          type: string
          description: User id that created the workflow definition.
        updatedBy:
          type: string
          description: User id that most recently updated the workflow definition.
      required:
        - name
...
    WorkflowNode:
      type: object
      additionalProperties: false
      required:
        - id
        - type
        - package
        - status
        - category
        - label
        - position
      properties:
        id:
          type: string
          format: uuid
          description: Node UUID
        type:
          type: string
          description: e.g. "playwright.open_page"
        role:
          type: string
          description: Execution role of the node.
          enum:
            - node
            - container
            - middleware
        package:
          $ref: '#/components/schemas/NodePackage'
        status:
          type: string
          description: Node lifecycle state.
          enum:
            - draft
            - published
            - deprecated
        category:
          type: string
          description: Group/category shown in the builder palette.
        label:
          type: string
        description:
          type: string
          description: Longer description of the node behaviour.
        tags:
          type: array
          description: Keywords for search/filter.
          items:
            type: string
        position:
          type: object
          required:
            - x
            - 'y'
          additionalProperties: false
          properties:
            x:
              type: number
            'y':
              type: number
        parameters:
          type: object
          description: Default parameter payload seeded from the manifest schema.
          additionalProperties: true
        results:
          type: object
          description: Default results payload seeded from the manifest schema.
          additionalProperties: true
        state:
          $ref: '#/components/schemas/WorkflowNodeState'
          description: Platform-managed execution hints surfaced in the builder (optional).
        schema:
          type: object
          description: JSON Schemas for parameters/results.
          additionalProperties: false
          properties:
            parameters:
              type: object
              additionalProperties: true
            results:
              type: object
              additionalProperties: true
        ui:
          $ref: '#/components/schemas/NodeUI'
        middlewares:
          type: array
          description: Ordered list of middleware definitions attached to this node.
          items:
            $ref: '#/components/schemas/WorkflowMiddleware'
...
    NodePackage:
      type: object
      additionalProperties: false
      required:
        - name
        - version
      properties:
        name:
          type: string
        version:
          type: string
...
    WorkflowEdge:
      type: object
      additionalProperties: false
      required:
        - id
        - source
        - target
      properties:
        id:
          type: string
        source:
          $ref: '#/components/schemas/EdgeEndpoint'
        target:
          $ref: '#/components/schemas/EdgeEndpoint'
...
    ResultError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
...